move the logic of the completion from lspClient to complete.ts
